+++ packages/terra-core-docs/src/terra-dev-site/test/cell-grid/CellGridDivider.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CellGrid, { Cell } from '../../../index';
+import CellGrid, { Cell } from 'terra-cell-grid';
 
 const CellGridDividerTest = () => (
   <div>