+++ packages/terra-cell-grid/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Migrate docs to terra-core-docs.
+
 ## 1.11.0 - (September 21, 2021)
 
 * Changed